Shuffle track class locations
